Change accessibility utils to support newer version of Select to Speak
Summary:
Android's "Select to Speak" service used to register for all AccessibilityEvent types, and would therefore pass the check in Litho's AccessibilityUtils class to be eligible for the more complex view hierarchy rendering needed for some accessibility services. It seems like with the latest version, they have paired down which events they register to only a handful, which was making it fail the check.

This changes the event looked for from "TYPE_VIEW_ACCESSIBILITY_FOCUSED" to "TYPE_WINDOW_CONTENT_CHANGED", which is included in the set that "Select to Speak" registers. The new event is the closest one to likely be necessary for *any* service that wants to parse the contents of the view hierarchy. It goes to reason that if you care about parsing that content, you also care about when it changes.

In theory, we could remove this check entirely, but this would cause many devices which have accessibility services enabled for non-accessibility purposes (password managers, virus scanners, etc.) to pass this check and get the more complex rendering unnecessarily.

Differential Revision: D21160590

fbshipit-source-id: 364bc191d592eecc621203a9f3e2cbc051b83e96