Remove assertMainThread() assertion from ComponentTree#getLithoView()
Summary:
The reason for this is that in tests, we are sometimes hitting ComponentTreeDumpingHelper (due to use of deprecated InlineLayoutSpec). ComponentTreeDumpingHelper gets the LithoView which in debug builds causes us to crash. In prod, it's likely we're calling getLithoView from a bg thread, but it doesn't crash since it's not a debug build.

The two options are to conditionally get the LithoView based on the thread (which doesn't feel great), or remove the assert. I'm choosing the latter because it seems valuable to keep in debug traces even if it's racy.

Reviewed By: adityasharat

Differential Revision: D21401777

fbshipit-source-id: 3356bec2cb972db0ed5927d03313fc1b5e883dcf