Crash in checkMainThreadLayoutStateForIncrementalMount only when incremental mount is enabled
Summary:
After D21229400 landed, we started calling notifyVisibleBoundsChanged even with incremental mount disabled. This means that now we're crashing when checking checkMainThreadLayoutStateForIncrementalMount in notifyVisibleBounds more often if `isLayoutRequested` is false.
Based on the crash message we only care about this if incremental mount is enabled, so not crashing otherwise. This should mitigate the crashes which started occuring after D21229400 landed - it doesn't fix the underlying issue which seems to be in the product.

Reviewed By: astreet

Differential Revision: D21400315

fbshipit-source-id: 957a695c0f881ad472385276a70b5e17ee18909d