Clean up DynamicPropsManager when item is unbound
Summary:
DynamicPropsManager is caching ComponentHosts. The cached entries are cleaned up when the LithoView is detached, but items which were already unmounted will not be removed from cache: https://fburl.com/diffusion/yn1wnrgf
This makes sure that the entry in DynamicPropsManager is always cleaned up when an item is unbound. Working on a test, just wanted to make sure this is correct first.

Reviewed By: pasqualeanatriello

Differential Revision: D21404023

fbshipit-source-id: fa44e9e4a2ce770bc4d5ca2086cf3842fc24971f