Refactor first mount flag.
Summary:
There was an error in the way we implemented how to check if it is the first mount of a tree. In the old MountState we would get a flag when a tree has been mounted and then we keep updating that. The new implementation would only check if a tree has been mounted but not if it is the first mount.

This change now adds that information to the component tree and we check the exact same thing as before.

Reviewed By: mihaelao

Differential Revision: D21306665

fbshipit-source-id: ca7926e693e118a06b8083101bbe915faf9448f5