Handle case where there are racing setRoot and setSizeSpec calls
Summary: You can have a case where you setRootAsync with a component, and in parallel a setSizeSpec that is compatible with the currently committed LayoutState. If those same new specs are not compatible with the new component's layout as well, we were never commiting an updated layout with the new root.

Reviewed By: adityasharat

Differential Revision: D21427615

fbshipit-source-id: fba232cd54933a8d6382f2f7693094c775a0fbeb