Introduce LifecycleTracker for tests
Summary:
I was looking into how can we get rid of TestDrawableComponent and found that we already have `MountSpecLifecycleTester` which could be replacement for that. But then I found some issues...

Currently `List<LifecycleStep.StepInfo> info` is used to track the lifecycle methods. As long as we are interested on particular sequence of events, that would work.

However, if we want to check current state like whether something is mounted or bound after multiple layouts it would be tricky to check that with only sequence of LifecycleSteps. Like if we want to check if smth is mounted just looking for ON_MOUNT is not enough as it might have been mounted first and then unmounted in which case it won't work as expected.

I'm proposing a new class - LifecycleTracker to track everything that happens to component both sequence of lifecycle events and current state (e.g. mounted, bound, attached).

adityasharat, colriot, what do you guys think about this approach?

adityasharat, also, why do you need to wrap LifecycleStep with StepInfo while you could use LifecycleStep directly: https://fburl.com/diffusion/wdw5cmar?

Reviewed By: adityasharat

Differential Revision: D21428993

fbshipit-source-id: 2ed3cbef684813c8243ed2c14059d600429c7595