Correctly process visibility outputs with extension when present
Summary:
LithoView can call processVisibilityOutputs on MountState. In that case, when MountState was using a visibility extension to process visibility outputs, the extension was not being used and events were processed with the old codepath.
This consolidates tests and ensures that the correct path is used - otherwise we hit a NPE.

Reviewed By: adityasharat

Differential Revision: D21548468

fbshipit-source-id: f33f2c3c88bb28718b86a71064d8e87f3b740f90