Subtracts the host translation to bounds before binding
Summary:
Subtracts the host translation to bounds before binding, for users of RenderCore this change will not make any difference. This change helps Litho interop with RenderCore. Litho uses absolute bounds in its LayoutOutput and then subtracts the host translation, this change allows for the exact same thing while using RenderCore.

i.e. When Litho is running on RenderCore, the bounds in the RenderTreeNode will be absolute and host translation will be properly set (not 0) (this is possible because Litho creates its own RenderTreeNode)

Reviewed By: mihaelao

Differential Revision: D21154669

fbshipit-source-id: 87983af73402f6f8ad985f59cfdc82a9c181d090