Don't use RenderCore's root render unit
Summary:
In order to apply the view attributes for the root host view when using Litho with RenderCore the LithoRenderUnit Binders need to be called for the root host view as well.

So, Litho cannot use RenderCore's root RenderUnit (`Reducer#sRootHostRenderUnit`), and this diff replaces that usage with a `LithoRenderUnit`.

Reviewed By: mihaelao

Differential Revision: D20869593

fbshipit-source-id: deaf89d4c52a24344828ba42cf52492f949d7c42