Overrides HostComponent's mount and bind methods to stop calls to super.
Summary:
This fixes an NPE thrown during mounting and binding the HostComponent when RenderCore's MountState is enabled. It works correctly in Litho because the ComponentContext is picked from the MountState when Component has a null Context on it (like in the case of the HostComponent).

By overriding the `mount` and `bind` methods of HostComponent we can bypass the problem of passing the ComponentContext to it via the LithoRenderUnit. Since the `HostComponent` doesn't use `ComponentContext` in its `onMount` or `onBind` methods this shouldn't be a problem.

Reviewed By: mihaelao

Differential Revision: D21130159

fbshipit-source-id: 2a7c46266a38c7c2a54c41b88e42f5ded1ac8d5a