Sets and unsets view attributes for LithoRenderUnit
Summary:
This diff wires the setting and unsetting of view attributes from the Litho RenderUnit Binders.

1. Make the set and unset code in MountState reuse-able for Litho and Litho on RenderCore.
1. Caches the default view attributes value flags in LithoRenderUnit to be used in unsetting view attributes.

When using RenderCore's MountState Litho does not create the MountItems and therefore does not create a LithoMountData which hosts the default view attribute value flags. To have feature parity in RenderCore mount the flags are hosted in  LithoRenderUnit.

Reviewed By: mihaelao

Differential Revision: D20623801

fbshipit-source-id: 8484295d3283f2d098a19dd8df560a4793c0f884