Fix litho view hierarchy dump
Summary:
The main issue is that the `LithoView` root was missing from the hierarchy as reported [here](https://fb.workplace.com/groups/jeste2e/permalink/2431400817151416/).
It was a major fuckup on my part as it resulted in strange bugs and subsequent hacky fixes.
By having the default view dump for the root we can use the proper bounds calculations.

Note on using `getBoundsInLithoView`:
By using the "absolute" bounds we can overcome the issue seen before where two elements would have the same bounds though they are in a hierarchy resulting in double offset. Using absolute and offseting from the parent fixes this problem.

Note on scrolling:
If the root LithoView is inside a scrollable container we won't have the correct position the same way we never have it with native dump. To fix it we need to dump the native view using custom dump flow we have for E2E. This will be handled in the next diff.

Reviewed By: scotthovestadt

Differential Revision: D21571657

fbshipit-source-id: b626795b560567dfd226d244197d598861b91749