Fix incorrect key generation after shallow copy
Summary: In shallow copy (which happens during state updates among other things), we were unsetting mHasManualKey which would cause us to generate different keys in the new layout if we had multiple of the same component under a parent. See the tests attached for an idea of what would trigger this in practice.

Reviewed By: adityasharat

Differential Revision: D21594900

fbshipit-source-id: f006840ca983d7490b200c0037a1c1fc6196bdd7