Pass localVisibleRect to HostListenerExtension callbacks
Summary:
Remove Host#getVisibleRect and pass visible rect directly to the HostListenerExtension callbacks.
There were some bugs coming from mPreviousMountVisibleRectBounds not being updated correctly before returning it in getVisibleRect - this makes it so that the extensions are passed the same localvisiblerect which is now passed into MountState and simplifies the codepaths where mPreviousMountVisibleRectBounds needs to be updated.

Reviewed By: adityasharat

Differential Revision: D21572973

fbshipit-source-id: 84a2c6e5017189cc646bc882110b768906022b50