Add test for VisibilityOutputsExtension crash
Summary:
I've fixed this crash before but it resurfaced after some refactoring. Adding a test to make sure that doesn't happen again.
tl;dr when the first dirty mount is triggered on a LithoView which is animating, the parent LithoView triggers an incremental mount on its children LithoViews but disables visibility processing. we still need to set the input on the visibility extension in that case to use it on the next mount

Reviewed By: adityasharat

Differential Revision: D21619333

fbshipit-source-id: 1ef3ba541f35cc23720b7cbdc8da2f0338ecd204