Add test that we always have an up-to-date main thread LayoutState at the end of measure
Summary:
This caused a UBN when landing the simplified CT threading model.

The basic issue is that before, measure was guaranteed to always set a main thread LayoutState, whereas now due to racing another thread could commit a valid LayoutState first. We need to make sure we promote that LayoutState even if we didn't compute it ourselves.

This adds a test for this case that fails in the original simplified threading diff.

Reviewed By: adityasharat

Differential Revision: D21659387

fbshipit-source-id: 05c7e365255445f3a862fe5eb2153e465cce1e7c