Fix NPE in VisibilityOutputsExtension when setting new ComponentTree
Summary:
If we call LithoView.setComponentTree multiple times on a LithoView without getting a chance to mount in between, the VisibilityExtension will not receive any input and will crash in onVisibleBoundsChanged.
Logically setting a new ComponentTree should not invoke an onVisibleBoundsChanged on the VisibilityExtension - this makes it explicit we want to clear all visibility outputs when a new componentTree is passed, same as the default behaviour.

Reviewed By: adityasharat

Differential Revision: D21661267

fbshipit-source-id: 20cf44dffab0ae2da108cd76d7d6a65883378913