Sets the layout state ids in the layout data for RenderTreeNode
Summary:
In order to reuse `MountState#shouldUpdateMountItem` in `LithoRenderUnit` the id of previously mounted `LayoutState` and the previous id of the next `LayoutState` is required.

Since `LithoRenderUnit` does not have access to this data, this diff stores these 2 ints inside a hash map and sets it on the layout data for the `RenderTreeNode` in `LayoutOutput`.

Reviewed By: mihaelao

Differential Revision: D21619087

fbshipit-source-id: 27d4157d8aa2ee4e9510ccbeb69ba17bde51259f