Adds an empty output when there are no mountable outputs when delegating to RenderCore
Summary:
Adds an empty output when there are no mountable outputs when delegating to RenderCore. Render expects at least 1 (the root) RenderTreeNode to be present even if there are no mount layout results. Litho on the other hand does not this constraint. This diff allows the interop to working correctly when there are no mountable outputs.

This will become unnecessary once the migration to RenderCore (layout+mount) is completed.

Reviewed By: mihaelao

Differential Revision: D21722497

fbshipit-source-id: 1bc5bb24d7367f46fff8c8ef2fd1a0eecf646e82