fixing PTR context problem in RecyclerSpec
Summary:
## Problem
When the UI triggers a PTR event on a Litho surface (with `RecyclerComponentSpec`), it'll dispatch a `PTRRefreshEvent` to its handler, which is defined as a Param in `RecyclerSpec`.

However, the handler instance in `RecyclerSpec` contains a snapshots of the State values that are potentially being outdated. The reason is that the Spec still uses an old instance from the `OnPrepare` stage, but in reality the instance with the latest Context has been assigned multiple times during onBind/onUnBind.

As a result, the states values are outdated, which causes a bug in Groups Engagement Tab.

See details: https://pxl.cl/17wmr

## Solution
Assign a new `PTRRefreshEvent handler` instance when onBind() is triggered. This will ensure that the handler will be able to pull the latest State values.

Reviewed By: muraziz

Differential Revision: D21696215

fbshipit-source-id: 9ac044acb739572c9451efd1ff76d751d465aeb9