Refactor Litho's implementation of touch expansion
Summary:
This change is in an effort to remove `Host` from the RenderUnit `Binder` APIs. This diff address the use of the Host for touch expansion delegation.

The way this diff now handles touch expansion delegation is by registering expansion against the MountItem instead of the touch expansion bounds.

This has an added benefit as well; the touch expansion does not need to be unregistered and registered again for every update of a mount item (saving some CPU cycles).

This diff also fixes a bug where the touch expansion was not unregistered in the following case:

1. Component created with touch expansion.
2. Component updated with touch expansion removed.
3. Unmount all items from the LithoView.

This has been captured in a new unit test as well.

Reviewed By: mihaelao

Differential Revision: D21765241

fbshipit-source-id: 3bc5ee1c7b1237ea4195e12b2d88c509f1055e56