Remove isAnimationLocked usage from IncrementalMountExtension
Summary:
There are two usages of isAnimationLocked inside IncrementalMountExtension, this gets rid of one of them.
When the TransitionsExtension is being used, this call is completely unnecessary in IncrementalMountExtension, since the transitions extension will increase ref count and will make an item mountable even if it's off screen.
However, we still needed this to test the IncrementalMountExtension inside MountState without a TransitionsExtension, when animations are still handled by MountState. This removes that dependency by moving that logic to the MountState and making sure we only use if and only if a TransitionsExtension is not present.

I'll send another diff to get rid of the other usage and remove this from the API entirely.

Reviewed By: adityasharat

Differential Revision: D21885385

fbshipit-source-id: 5fbbdf314dd299c0d02cdd174842f561d55fc16c