Remove mountViewIncrementally from IncrementalMountExtension on animation locked
Summary:
IncrementalMountExtension should not have access to information about animations.
When an item is locked for animation and it has children LithoViews, we need to ensure that all its children LithoViews are mounted even if incremental mount is enabled and the LithoViews are not on screen.
If transitions are handled by Litho MountState, we ensure the children LithoViews are mounted from MountState. If there's a TransitionsExtension, we delegate to the transitions binder to ensure mounting when the parent is mounted.

Reviewed By: adityasharat

Differential Revision: D21929429

fbshipit-source-id: 09af9e9b3633df8354f6e96ac64af112292b5669