Remove TransitionExtension field from MountDelegate
Summary:
A bunch of tests are failing in D21885385 because of the new checks added around calling isAnimationLocked when a TransitionsExtension is present. That's because we were relying on the mTransitionsExtension field in MountDelegate to assert it exists, but that field was actually never written to even if a TransitionsExtension is registered on the MountDelegate.
This removes that unused field and adds a temp workaround to call isAnimationLocked on the extension, until I send the diff to remove that as well.

Reviewed By: nicous

Differential Revision: D21903728

fbshipit-source-id: 35ff86d5cc501a7877ee02539ce5a295fe0ca61c