Process visibility events when parent is in transient state
Summary:
I misunderstood how the transient state is interpreted right now in LithoView. When the view is in transient state, we disable processing visibility events - it has nothing to do with the transient state of the parent.
This was causing an issue when using the visibility extension because some events were not processed correctly if the parent of the LithoView was in transient state. Added a test too.

Reviewed By: adityasharat

Differential Revision: D21974323

fbshipit-source-id: bc2fb753b68226ac325c6a075b48975f29dc46b9