Introduce RenderTreeHost & extract rendering logic from RenderState
Summary:
# Problem
We want to remove `RenderState` from sub-`RenderTree`, to keep only one `RenderState` at the root of the tree. This means we are no longer keeping `RenderState` around for each children in a collection.

So we want to be able to measure/layout/render a `LazyTree` without `RenderState`. All of that currently live in `RenderState#RenderResultFuture`.

# Solution
Decouple and refactor the code for sharing between different files.

- Extract the logic into a helper class `NodeResolveUtil`, and use it in `RenderState`.
- Create a new host `RenderTreeHost` to differentiate it from `RootHost`.
- Implement a `HostView` with `RenderTreeHost`.
- Move `RenderResult` out of `RenderState` namespace.

Reviewed By: pasqualeanatriello

Differential Revision: D21747661

fbshipit-source-id: 4edfc7b7830bd79073135f1cbd273cca1195370b