Support UNSPECIFIED and AT_MOST measurements in CollectionNode
Summary:
# Problem
In D21503209, I tried to support wrap_content in `CollectionNode` by getting a direct access to the `RenderState` on the viewport so that I can get my size based on their measurement.

The approach was to reuse those measurements. After speaking to pasqualeanatriello, we think that we should avoid accessing `RenderState`, and instead treat each of these children as part of a big RenderTree.

Things we want:
- Hide `RenderState`, there ideally should only be 1 RenderState for the entire Bloks tree
- Move layout/measure from `RenderUnit` to `NodeComponent`
- Reuse `RenderResult` whenever possible

# Solution
`RenderState` is completely removed from `SimpleListAdapter`. In its place, we have a map of item -> `RenderResult`.

Those `RenderResult` will be computed either during `layout()` or in `onBindViewHolder()`.

Reviewed By: pasqualeanatriello

Differential Revision: D21572916

fbshipit-source-id: 6295daf9bd23a057e53aafbdd532f8ce41a0bbb3