Replace TestDrawableComponent with SimpleMountSpecTesterSpec and MountSpecLifecycleTesterSpec
Summary:
Working towards the goal of deprecating TestDrawable component, this diff goes through all the usages in `litho-it` and replaces them with either `SimpleMountSpecTesterSpec` or `MountSpecLifecycleTesterSpec`.

The former is just a UI block with size and a background for layout and diffing purposes. `MountSpecLifecycleTesterSpec` adds a lifecycle tracker to it for other assertions.

Reviewed By: adityasharat

Differential Revision: D21904082

fbshipit-source-id: 78c28468906750789e138b26659967ab949e9119