Replace TestDrawableComponent with MountSpecLifecycleTesterSpec in litho-widget
Summary:
Working towards the goal of deprecating TestDrawable component, this diff goes through all the usages in `litho-widget` and replaces them with either `SimpleMountSpecTesterSpec` or `MountSpecLifecycleTesterSpec`.

The former is just a UI block with size and a background for layout and diffing purposes. `MountSpecLifecycleTesterSpec` adds a lifecycle tracker to it for other assertions.

Reviewed By: muraziz

Differential Revision: D21976667

fbshipit-source-id: 2a7c6fde658e3e562f92b5c1de8d1d0ee96facdc