Use SimpleMountSpecTesterSpec in litho-it-powermock
Summary:
Working towards the goal of deprecating TestDrawable component, this diff goes through all the usages in `litho-it-powermock` and replaces them with either `SimpleMountSpecTesterSpec` or `MountSpecLifecycleTesterSpec`.

The former is just a UI block with size and a background for layout and diffing purposes. `MountSpecLifecycleTesterSpec` adds a lifecycle tracker to it for other assertions.

Reviewed By: colriot

Differential Revision: D21975390

fbshipit-source-id: 1f829881abd0aa3cd3a83a1fa5ff8b1cf13c7199