Refactors A11Y invalidation
Summary:
Adds new property `mImplementsVirtualViews` to `HostComponent` and  `ComponentHost`. `HostComponent#mImplementsVirtualViews` will be set to `true` if and only if any of its child's outputs are:

* Accessible: `#isAccessible()`
* Implement extra A11Y nodes: `Component#implementsExtraAccessibilityNodes()`

This is done retroactively by setting `mImplementsVirtualViews` on the parent output (which should always be by definition a `HostComponent`) just before adding an output it in `LayoutState#addMountableOutput()`.

This caches the result of `ComponentHost#implementsVirtualViews()` in the `HostComponent` and eventually the `Host` (in `OnBind`). The current implementation infers `ImplementsVirtualViews` by computing it for every mounting/unmounting/updating item. This will possibly be more performant as well.

This is also possibly less buggy because the current implementation of `ComponentHost#implementsVirtualViews()` returns `true` if and only if the **first accessible item** implements **extra accessibility nodes**, which might result in missing calls to invalidate A11Y state (say if the second accessible item implements extra accessibility nodes).

Reviewed By: mihaelao

Differential Revision: D22042887

fbshipit-source-id: 92d59dea490b456ae1393fa55d31681b3f3f417c