Introducing DerivedDynamicValue
Summary:
Introducing DerivedDynamicValue, which allows a new DynamicValue to be derived from another i.e. wrap it and apply some modifier to it.

This is useful as it allows a single value to be passed to a component (e.g. a single Float representing animation progress), and all the internal DynamicProps can be derived from that (e.g. alpha, translation, rotation).

Reviewed By: astreet

Differential Revision: D22097645

fbshipit-source-id: d731c09ca9fb27a8e8fc47e8c7af3b123c7f1445