Fix NPE when reusing LithoViews
Summary:
When reusing LithoViews in very specific scenarios we could end up getting a NPE with the mTransitionManager not being initialized.
The problem was that we check if the component tree is different to assert if we should animate or not. We wrongly had the property mLastMountedComponentTreeId updated in beforeMount.
This code was ported from Litho MountState.java and there we were updating the mLastMountedComponentTreeId at the end of the mount method.

This is the Logview of the error being fixed here: https://www.internalfb.com/intern/logview/details/facebook_android_crashes/1a8f533e7a15f8df23f8d0c4e4c6749d/?trace_key=4d6b5326cd7e951f61eff562d6f68795&tab=All

Reviewed By: mihaelao

Differential Revision: D22119037

fbshipit-source-id: fe8417b6d8cafa4bd80884e109a62c391ecc4a02