Uses Host#unmount(MountItem) to avoid index mismatch
Summary:
Uses Host#unmount(MountItem) to avoid index mismatch.

Using the index from the layout output is not always correct; for example when disappearing animations move the item to the root the actual mount index would change (i.e. mismatch from the one in its layout output).

To avoid this, it is far safer to call the `Host#unmount(MountItem)` which does not suggest an index; the current implementation finds the current index of the item in Host.

Reviewed By: nicous

Differential Revision: D22159273

fbshipit-source-id: 707405319e3a06912d4a8b0a1e0f1fb92f39a40c