Refactors implementation of Drawable state in ComponentHost
Summary:
Refactors implementation of Drawable state in ComponentHost.

Uses the presence of `LithoMountData` on the `MountItem` to infer if render core is enabled or disable. This way avoids the excess call to set the drawable state when render core is enabled (once in the ComponentHost, and then again in the Binder).

Reviewed By: mihaelao

Differential Revision: D22136718

fbshipit-source-id: 4ff5aac29d4ae3d0d09f9625d7ef1f614789cf9a