Reset root view attributes after unmountAll
Summary:
After calling unmountAll, `maybeUnsetViewAttributes` is called for each item that gets unmounted.
We want to make sure that after an item is remounted after `unmountAll`, the view attributes are set correctly.

The root host doesn't get removed from the list of mounted items when `unmountAll` is called, but `maybeUnsetViewAttributes` is still called on it. On the next mount, since the root host is present in the list of mounted items, Litho was skipping it to set the view attributes again. This diff fixes that case and makes sure the view attributes are reset on the host after `unmountAll` is called.
See the test for an example - it was failing prior to this change.

Reviewed By: adityasharat

Differential Revision: D22115239

fbshipit-source-id: 09f9ec91cc49060197641b82f69c1e7b8645e27a