Call mount if LithoView called unmountAllItems
Summary:
When using a MountDelegateTarget in LithoView, we only call the MountDelegateTarget's mount method when a dirty mount occurs. Otherwise, we only notify the extensions that the LithoView's visible bounds changed (this is the equivalent of calling a non-dirty mount in Litho).
After LithoView performs an `unmountAll`, it puts the MountState into a state where `mNeedsRemount` is true and it flips this flag after the next mount occurs. However, when using a MountDelegateTarget, if unmountAll is followed by a non-dirty mount, that flag will not get flipped until the next dirty mount because we don't call the MountState's mount.

This adds a test to make sure that flag is flipped correctly and makes sure that `unmountAll` is followed by a mount call on the MountDelegateTarget.

Reviewed By: adityasharat

Differential Revision: D22115127

fbshipit-source-id: 6b4bfefcd40d66a0b90b28eeb6b19899f7120a97