Add test for mounting component with negative margin
Summary:
The existing test was not working anymore, because it ended up with the host always having the same bounds as the child. This fixes the test so that if we remove these lines: https://fburl.com/diffusion/6s7pgs2t the test fails.
Those lines are necessary to account for the case when a layoutOutput's bounds are outside of its host. For incremental mount, the mountable items are sorted by their tops and bottoms, so there are cases when a child can be mounted before its host if they were ordered in such a way when the child can have negative margins for example.

Reviewed By: adityasharat

Differential Revision: D22137133

fbshipit-source-id: 082d50afcbc67f0214a8e72a212523c269766ab6