Fixes bug with TextInput TextChangedEvent
Summary:
This diff fixes a bug where if a on text change event calls the `TextInput.setText` lifecycle, that update is never applied.

This was happening because the `TextInput` is saving the text (`EditTextWithEventHandlers#mTextState`) after dispatching the `TextChangedEvent`. So changes made in the event handler were being overridden.

This diff fixes this by saving the current text before dispatching the event.

Reviewed By: Katalune

Differential Revision: D22185274

fbshipit-source-id: 84480c23fb3dee215b16b3eab3b4c87306c3b4b1