Prevent NPE in VisibilityOutputsExtension
Summary: It's possible that the LithoView will notify the VisibilityOutputsExtension that the visibility bounds might have changed before the LithoView was ever mounted, which means we didn't get a chance to call beforeMount and set the input for the extension. That can happen for example when calling setVisibilityHint on a LithoView with children (which will call setVisibilityHint recursively on its LithoView children) - for the LithoView children which were not already mounted this will crash, but it should be a no-op instead.

Reviewed By: nicous

Differential Revision: D22189741

fbshipit-source-id: c5b0eda22d912a34cba31642fde4eb1eb3e60fcf