Add root output of RenderTree to mOutputsIdToPositionMap
Summary:
There is a crash spiking when we turn on using a MountDelegateTarget in LithoView, caused by an IOB exception.
The incremental mount extension is iterating over a list of mountable items from LayoutState and  trying to find the position of the item in the list of LayoutState's mountable items by calling getLayoutOutputPositionForId: https://fburl.com/diffusion/immifw4i. Calling this method returns -1 even if the item is present in the list of mountable items, so it's missing from mOutputsIdToPositionMap.
This is the one place where we don't add the LayoutOutput to mOutputsIdToPositionMap when we add to the list of mountable items

Reviewed By: adityasharat

Differential Revision: D22209288

fbshipit-source-id: 0ce7dcb805edcb3da7d0486ec11da8b60fd4983e