Process visibility outputs after mount
Summary:
Right now in Litho visibility events are processed after mount.
This is different from what we were doing with a MountDelegate - the VisibilityOutputsExtension was processing events in the `beforeMount` call.
The issue is that by doing this, we are calling into user code before mount happens, so that means that mount will happen later. The other issue is that this potentially breaks visibility events which were handling mounted views, such as anchoring a tooltip to a view in a visibility callback.

This diff ensures the order that we have in Litho right now, so visibility events are always processed after mount happens.
I want to explore a different version to see if it's possible to process visibility with binders instead of having a VisibilityOutputsExtension.

Reviewed By: adityasharat

Differential Revision: D22232497

fbshipit-source-id: be8f2c286c9ed0bb537fe3d0092dca2e725538c5