Fixes bug with ComponentHost padding when drawable outputs are disabled
Summary:
when background outputs are disabled, they are wrapped by a `ComponentHost`. A background can set the padding of a view, but `ComponentHost` should not have any padding because the layout calculation has already accounted for padding by translating the bounds of its children.

This fixes it the issue by manually setting the padding to zero only for hosts when the config is enabled.

Reviewed By: muraziz

Differential Revision: D22259244

fbshipit-source-id: 5a6a6c0c24885c13d795ee3c90703f73bfbab6f8