Reset dynamic props on unbind
Summary:
DynamicProp values set on a mounted content need to be reset when the content gets unbounded.
This is similar to how view attributes are unset when an item gets unbound. However, when dynamic prop values are set on a mounted content, the NodeInfo does not get mutated, so unsetViewAttributes does not clear those values in onUnbind. DynamicPropsManager needs to handle that explicitly.

Here's what can happen if the values are not reset. Consider the following mount hierarchy:

```
- HostComponentRoot: id 0
|
|------ HostComponentA (id: 1, ComponentHostA)
|          |
|          | -- ComponentA: dynamicValue
|
|
|-------HostComponentB (id: 2, ComponentHostB)
|          |
|          |-- ComponentB

```

ComponentHostA will be applied the dynamic prop values specified by ComponentA.
After a state update, ComponentA gets removed from the hierarchy and we end up mounting this:

```
- HostComponentRoot: id 0
|
|------ HostComponentC (id: 1, ComponentHostA)
|          |
|          | -- ComponentB
```

HostComponentC will be assigned the same layout output id as HostComponentA, because the layout output ids are generated based on the position of a node in the tree. MountState will find that there's already a mounted item for this layout output id, with ComponentHostA as content, and will binding HostComponentC to ComponentHostA. If dynamic prop values aren't reset when rebinding, then ComponentHostA will be assigned dynamic prop values from the first layout, even if the new component doesn't have them.

This is what happens in the bug in t68986811, where the ComponentHost keeps the alpha value from a previous binding and it gets reused with a Component which was not transparent.

Reviewed By: adityasharat

Differential Revision: D22410906

fbshipit-source-id: 689da1596e50a2464bedce62b0b00b55f45143e5