Remove disappearing mount content when unmounting or changing tree
Summary:
When we unmount an item that's a host or remounting a host for a new component tree we need to remove disappearing mount content. This diff allow for this to work when we are using the transitions extension.
I need to move the previous removeDisappearingMountContentFromComponentHost within unmountItem to unbindMountItem as we don't want to call that on items that should delegate mount.

Reviewed By: adityasharat

Differential Revision: D22358650

fbshipit-source-id: 9b428137c846816b2cd688f41efe990e7ba3c12d