use reflection to dump litho view
Summary:
To have better control over the litho dump we should be able to call it directly instead of via toString of the litho view.

It's required to pass the extra properties flag but it's also nice for not having to calculate depth manually and removing the native view from the toString.

I don't want any direct or shared dependency because I want to inject this code in Android OS image.

Reviewed By: scotthovestadt

Differential Revision: D22434845

fbshipit-source-id: ae1353951714dbe4e95fa4e4781c2db12e7f609a