Add editor data format
Summary:
This diff adds a format for structural types that's meant to be convertible to any communication protocol with editors such as Flipper.

It is a sealed class so much as Java allows. `when` is used to operate on the blessed inheritors, and `aggregate` does it with a carry value. Then `whenPrimitives` is a special case made to easily find "leaf" nodes down a deeply nested shape. We'll use these on the stack to implement features ;)

I've called it `shape` instead of `object` because it's already a heavily overloaded term, and `shape` is what's used in Hack.

Reviewed By: muraziz

Differential Revision: D22437766

fbshipit-source-id: d30f5f6f13cb68646e01b8a79e373924a885a08d