Add Editor, EditorRegistry and basic Instances
Summary:
This diff adds the concept of an `Editor`, a class capable of operating on a Component field to understand its structure and apply changes to it.

It comes with an `EditorRegistry` that maps an `Editor` instance to the `Class` it is implemented for. Sadly this cannot be enforced safely in Java, so we need to use a map of `Class<?>` instead.

The diff comes with a few pre-written instances meant to replace some of the existing code for basic types. These instances, and I suspect many more, will need to operate on fields using reflection in an UNSAFE way, so I added helpers for it marking them clearly as such.

Reviewed By: muraziz

Differential Revision: D22333048

fbshipit-source-id: 14a97babfa011326e6847282179f2ac715b0d556