Fix crash on LithoView xy transition
Summary:
On every frame of a LithoView x,y transition we were repeatedly setting a value and resetting it to 0. This results in a crash. This was due to attempting an incremental mount after each change to the view's x,y coordinates. Which includes a call to AnimationProperty.reset(). We don't need an incremental mount because Animating outputs are already fully mounted.

NB: This diff is fixing an existing issue that became more pronounced with the introduction of D22390161
NB2: This is not an issue with RenderCore due to interaction of TransitionsExtension and IncrementalMountExtension by refcounting.

Reviewed By: nicous

Differential Revision: D22432289

fbshipit-source-id: 266abbd292286db9fc50de1d017c28192c6616f1