itho-intellij-plugin][2/n] Refactor: Use ComponentGenerateUtils#updateComponent instead of componentsCache#maybeUpdate
Summary:
Currently we have 2 ways to update component. We want to simplify logic by leaving one way only. This Diff replaces usage, so that the method could be deleted in next Diffs.

Keep: ComponentGenerateUtils#updateComponent which looks for in-memory or on-disk component, and updates if it's present.
Remove: ComponentsCache#maybeUpdate which updates component if it's absent.

Reviewed By: cute-jumper

Differential Revision: D22459642

fbshipit-source-id: c3df5ac0c70a9cac70b3fd51a988185c60970a7d