Remove unnecessary runtime exception.
Summary:
When we unmount an animating item, we relay on a WeakReference on that mount item being deallocated to stop updating the view/drawable. While triggering mutiple tree updates it happens that the reference can be found just before the WeakReference deallocates it.
Removing this runtime exceptions would only help by not crashing the app and allow for the reference to be removed.

Reviewed By: andrewpmsmith

Differential Revision: D22233928

fbshipit-source-id: 2360b6d7d75653a46fea5b8092918c613da545e6