Fix ClassCastException in e2e tests with MountDelegate enabled
Summary:
Fixes a ClassCastException in EndToEndTestingExtension, when trying to cast a ComponentHost to a com.facebook.litho.Host.
The original diff intended a cast to com.facebook.rendercore.Host which is still not correct because it's used with a litho TestItem which expects a ComponentHost. Not sure why this was not an issue when it was first introduced.

Reviewed By: adityasharat

Differential Revision: D22503429

fbshipit-source-id: 9aeec97b701b035b21f8571b9d94e2b35c237e7c